Исправил замечания наставника
